+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Notable changes to the Indicia warehouse are documented here.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Version 9.21.0
+*2026-03-23*
+
+* Adds an encrypt flag that can be used to enable encrypted storage of text custom attribute
+  values, e.g. email addresses. See https://github.com/Indicia-Team/warehouse/issues/581.
+  Information on required setup steps is given in the [Indicia installation guide](https://indicia-docs.readthedocs.io/en/latest/administrating/warehouse/warehouse-installation.html#encrypted-custom-attribute-values).
+
 ## Version 9.20.0
 *2026-03-18*